refactor(tui): relocate to src/tui/, split ChatScreen, extract usePolling/readSSE (#731)
Closes #731. Moves TUI out of legacy/ and untangles the architecture.

Relocate:
- src/legacy/tui/ → src/tui/
- Update cli.ts dynamic import + CLAUDE.md layout

ChatScreen split (445 → 123 LOC):
- messages.ts — pure historyToChatMessages, extractResultText, STEER_PREFIX const
- MessageView.tsx — renderMessage as a real component
- useChatSession.ts — encapsulates init/send/steer/abort/resetOwned with mode
  branching internally; ChatScreen no longer sees owned vs attach
- ChatScreen.tsx — thin composition (header / Static settled / active / input)

Shared infrastructure:
- usePolling.ts — used by StatusScreen + SessionsScreen (was duplicated)
- api.ts readSSE — extracted from sendMessage + attachStream (deduped parser)

Quality fixes:
- StatusScreen: process.exit(0) → useApp().exit() (was bypassing ink cleanup)
- api.ts: parseInt NaN guard for ISOTOPES_PORT
- getHistory return type now uses HistoryItem (was untyped inline shape)
- useChatSession sendOwned: build new block objects instead of mutating in place
- ChatScreen settled list: useState + queueMicrotask instead of mutating ref during render
- Comment explaining the React `key=` remount pattern in useChatSession

Net +79 LOC across the directory; the increase pays for clean module boundaries.
ChatScreen alone shrank by 322 LOC.
