+++ src/tui/commands.test.ts
