+++ src/tui/commands.ts
